[88] improve solution
Use predefined constant.
Lose unnecessary try, finally block. The function being decorated by
`contextmaganer` does only need to yield. At this point everything
that is in the with statements happens and afterwars at this point
we resume.
Try, except and finally block would only be useful for handeling
uncovered exceptions in the with block.